backend/requirements.txt
Flask
Flask-CORS
torch
torchvision
Pillow
numpy
opencv-python
gunicorn
wget
transformers
accelerate
protobuf